build: migrate to otterbrix 1.0.0b2-rc-1 (arrow 24, boost 1.88, grpc 1.69, catch2 v3) + kafka b2 port
Dependency-migration half of the former #22 (Spark Connect frontend is
stacked on top of this branch as PR #22).

Deps (conanfile.py): otterbrix a13-rc-3 -> 1.0.0b2-rc-1#1843c85 (revision-
pinned: resource_tracer reclaim-on-destroy fix), arrow 21 -> 24 (C++20),
boost 1.87 -> 1.88 (forced by otterbrix b1+; CMP0167 NEW), grpc 1.50 -> 1.69
+ abseil 20250127.0 override (b2 recipe abseil range; grpc 1.50 can't build
against abseil 2025+), catch2 2.13.7 -> 3.15.2 (b2 requires
find_package(Catch2 3)); Dockerfile.test protobuf/re2 cppstd=17 pins.

Source migration (a13 -> b2 API):
- multi-chunk cursor: cursor->chunk_data() -> cursor->chunks();
  session_payload.chunk -> chunks (pmr vector, <=1024 rows per chunk) with
  spanning size()/column_count()/empty(); all frontends iterate chunks
  (flight batch_reader streams one RecordBatch per chunk; PG row counts
  widened to int64); chunk_to_{csv,ndjson,parquet} vector overloads
  (COPY ... TO multi-chunk); drop unused utility/shared_flight_data.hpp
- flat plan nodes: node_drop_collection/index -> node_drop_t +
  drop_target_kind; node_catalog_resolve_t + resolve_kind
- LIMIT/OFFSET pushdown: generate_select dropped the node_limit_t child —
  every remote LIMIT n returned all rows; + 3 regression tests
- tsan: connector errors marshaled as query_outcome<T> values across the
  io->consumer boundary (asio use_future exception_ptr race, boost 1.88);
  keeps kafka's draining QueryHandleWaiter dtor
- api_server: drop stale duplicate *_connection_config.hpp headers (same-TU
  redefinition introduced by #23) incl. their CMake source entries

Kafka b2 port (kafka #24 merged on main as a13 code):
- 4 cursor->chunk_data() sites -> chunks()/value(col,row); >1024-row
  produce/stream/recover paths no longer silently truncate
- kafka_manager_ddl: make_node_drop(resource, drop_target_kind::collection)
- kafka_node: node_catalog_resolve_table_t -> node_catalog_resolve_t(kind)
- kafka_reader chunk_to_json: guard alias() (b1+ null-derefs on alias-less
  types), colN fallback fails the round-trip instead of segfaulting
- kafka tests to catch2 v3; discovery reconciled to otterstax_discover_tests

Host-macOS build enablement (ld64/dyld):
- LINK_GROUP:RESCAN and --whole-archive are GNU-ld-only -> APPLE guards
  (ld64 re-searches archives; -force_load for otterbrix_local)
- INSTALL_RPATH as semicolon list (colon string becomes one bogus LC_RPATH
  on macOS) + @loader_path twins for every $ORIGIN entry; ladder deepened
  to the kafka grammar test dir

New regression test: 2500 rows across >=3 chunks survive the scheduler.

+++ frontend/postgres_server/packet/packet_utils.cpp
@@ -20,7 +20,7 @@ namespace frontend::postgres {
     command_complete_tag::command_complete_tag(std::string tag)
         : tag(std::move(tag)) {}
 
-    command_complete_tag command_complete_tag::simple_command(NodeTag node, int32_t rows) {
+    command_complete_tag command_complete_tag::simple_command(NodeTag node, int64_t rows) {
         switch (node) {
             case T_CreateStmt:
                 return {"CREATE TABLE"};
@@ -46,10 +46,10 @@ namespace frontend::postgres {
                 return {"COMMAND"};
         }
     }
-    command_complete_tag command_complete_tag::select(int32_t rows) { return {"SELECT " + std::to_string(rows)}; }
-    command_complete_tag command_complete_tag::insert(int32_t rows) { return {"INSERT 0 " + std::to_string(rows)}; }
-    command_complete_tag command_complete_tag::update(int32_t rows) { return {"UPDATE " + std::to_string(rows)}; }
-    command_complete_tag command_complete_tag::delete_rows(int32_t rows) { return {"DELETE " + std::to_string(rows)}; }
+    command_complete_tag command_complete_tag::select(int64_t rows) { return {"SELECT " + std::to_string(rows)}; }
+    command_complete_tag command_complete_tag::insert(int64_t rows) { return {"INSERT 0 " + std::to_string(rows)}; }
+    command_complete_tag command_complete_tag::update(int64_t rows) { return {"UPDATE " + std::to_string(rows)}; }
+    command_complete_tag command_complete_tag::delete_rows(int64_t rows) { return {"DELETE " + std::to_string(rows)}; }
     command_complete_tag command_complete_tag::begin() { return {"BEGIN"}; }
     command_complete_tag command_complete_tag::commit() { return {"COMMIT"}; }
     command_complete_tag command_complete_tag::rollback() { return {"ROLLBACK"}; }
